refactored strategy options

+++ src/main/kotlin/dev/crashteam/repricer/service/KeShopItemStrategyService.kt
@@ -26,7 +26,7 @@ class KeShopItemStrategyService(
     }
 
     @Transactional
-    fun updateStrategy(shopItemId: UUID, patchStrategy: PatchStrategy): Int {
+    fun updateStrategy(shopItemId: UUID, patchStrategy: PatchStrategy): Long {
         return strategyRepository.update(shopItemId, patchStrategy)
     }
